+++ src/extractors.js
@@ -111,13 +111,11 @@ const yelpBusinessInfo = (json) => {
     };
 };
 
-const yelpBusinessReviews = (url, json) => {
+const yelpBusinessReviews = ({ url, json, scrapeReviewerName, scrapeReviewerUrl }) => {
     const reviews = new Map();
-
     json.reviews.forEach((review) => {
         if (!reviews.has(review.id)) {
             const $ = load(review.comment.text.replace(/<br>/g, '\n'), { decodeEntities: true, normalizeWhitespace: true });
-
             reviews.set(review.id, {
                 date: new Date(review.localizedDate).toISOString(),
                 rating: review.rating,
@@ -126,7 +124,15 @@ const yelpBusinessReviews = (url, json) => {
                     .get()
                     .filter((s) => s)
                     .join('\n'),
+                language: review.comment.language,
+                isFunnyCount: review.feedback.counts.funny,
+                isUsefulCount: review.feedback.counts.useful,
+                isCoolCount: review.feedback.counts.cool,
                 photoUrls: review.photos.map(((photo) => new URL(photo.src, url).toString().replace(/\/[^/]+.jpg/, '/o.jpg'))),
+                reviewerName: scrapeReviewerName ? review.user.markupDisplayName : null,
+                reviewerUrl: scrapeReviewerUrl ? `https://www.yelp.com/${review.user.userUrl}` : null,
+                reviewerReviewCount: review.user.reviewCount,
+                reviewerLocation: review.user.displayLocation,
             });
         }
     });